+++ test/unit_test/test_core/test_forwardChain.cpp
@@ -211,7 +211,7 @@ namespace original {
     }
 
     TEST(forwardChainTest, pushTest) {
-        constexpr int64_t dataSize = 100000;  // 大数据量测试
+        constexpr integer dataSize = 100000;  // 大数据量测试
         std::random_device rd;
         std::mt19937 gen(rd());
 
@@ -219,7 +219,7 @@ namespace original {
         std::list<size_t> l;
         for (int i = 0; i < dataSize; ++i) {
             std::uniform_int_distribution dist(0, i);
-            const int64_t rand = dist(gen);
+            const integer rand = dist(gen);
             auto li = l.begin();
             for (size_t j = 0; j < rand; ++j) {
                 ++li;
@@ -261,7 +261,7 @@ namespace original {
     }
 
     TEST(forwardChainTest, popTest) {
-        constexpr int64_t dataSize = 100000;  // 大数据量测试
+        constexpr integer dataSize = 100000;  // 大数据量测试
         std::random_device rd;
         std::mt19937 gen(rd());
 
@@ -269,7 +269,7 @@ namespace original {
         std::list<size_t> l;
         for (int i = 0; i < dataSize; ++i) {
             std::uniform_int_distribution dist(0, i);
-            const int64_t rand = dist(gen);
+            const integer rand = dist(gen);
             auto li = l.begin();
             for (size_t j = 0; j < rand; ++j) {
                 ++li;
@@ -279,7 +279,7 @@ namespace original {
         }
         while (!c.empty()) {
             std::uniform_int_distribution dist(0, static_cast<int>(c.size() - 1));
-            const int64_t rand = dist(gen);
+            const integer rand = dist(gen);
             auto li = l.begin();
             for (size_t j = 0; j < rand; ++j) {
                 ++li;
